refactor: issue #226
