+++ pyproject.toml
@@ -33,6 +33,8 @@ authors = [{ name = "0x676e67", email = "gngppz@gmail.com" }]
 keywords = ["http", "client", "websocket", "ja3", "ja4"]
 
 [tool.maturin]
+python-source = "python"
+module-name = "rnet._rnet"
 features = ["pyo3/extension-module", "pyo3/generate-import-lib"]
 strip = true
 
@@ -43,8 +45,8 @@ Repository = "https://github.com/0x676e67/rnet"
 
 [dependency-groups]
 dev = [
-  "pytest>=8.3.5; python_version >= '3.8'",
-  "pytest>=7.4.4; python_version == '3.7'",
-  "pytest-asyncio>=0.25.3; python_version >= '3.9'",
-  "pytest-rerunfailures>=15.0; python_version >= '3.9'",
+    "pytest>=8.3.5; python_version >= '3.8'",
+    "pytest>=7.4.4; python_version == '3.7'",
+    "pytest-asyncio>=0.25.3; python_version >= '3.9'",
+    "pytest-rerunfailures>=15.0; python_version >= '3.9'",
 ]
